back-end/requirements.txt
# Django and REST Framework
Django>=5.2.0
djangorestframework>=3.15.0
drf-yasg>=1.21.0
django-cors-headers>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# NumPy for vector operations
numpy>=1.24.0
